refactor: 💡 data feching for claims
